aspire: Set `allowInsecure` to true for internal services
In addition to setting `allowInscure` for HTTP2 (i.e. gRPC) services,
we now also set this to true by default for any service that is not
publicly exposed.

This allows inscure traffic to internal nodes, to allow requests to
things like `http://backend` from one internal service to another.

Fixes #3201