aspire: Add support for `dockerfile.v0` (#3075)
As part of Preview 2, Aspire is adding support for a new resource,
`dockerfile.v0`. This represents an image that should be built (from a
Dockerfile) and deployed.

Consider the following AppHost:

```csharp
var builder = DistributedApplication.CreateBuilder(args);

builder.AddNodeApp("nodeapp", "../NodeApp/index.js")
    .WithServiceBinding(3000, scheme: "http", env: "PORT")
    .AsDockerfileInManifest();

builder.Build().Run();
```

The resulting manifest looks like this:

```json
{
  "resources": {
    "nodeapp": {
      "type": "dockerfile.v0",
      "path": "../NodeApp/Dockerfile",
      "context": "../NodeApp",
      "env": {
        "NODE_ENV": "development",
        "PORT": "{nodeapp.bindings.http.port}"
      },
      "bindings": {
        "http": {
          "scheme": "http",
          "protocol": "tcp",
          "transport": "http",
          "containerPort": 3000
        }
      }
    }
  }
}
```

From a high level, these are very similar to `project.v0`
resources. The image needs to be built and pushed (either via `dotnet
publish` or `docker build` + `docker push`) and then we need to create
an Azure Container Apps App from it.

When doing `azd infra synth` we write the manifests for these types of
services in a `manifests` folder that is SxS with the
`Dockerfile`. This mimics the strategy for projects in spirit (where
we use the `.csproj` file as the "project" file)