Enable admin user for registry in ACA sample
This enables the admin user on the provisioned container registry so
that `azd` can connect to it to push even if RBAC is not working
because the user is not anowner of the resource.  We've seen cases
where users who's accounts where created a long while back are set as
[classic adminisrators](https://learn.microsoft.com/en-us/azure/role-based-access-control/classic-administrators)
and not Owners on their accounts and ther is presently a service issue
that prevents these users from exchanging their AAD creds for an
access token.

Fixes #3157