Add RPC Server for VS Aspire Deploy
This change adds a new command to `azd`, `azd vs-server`.  When run,
`azd` starts a listener on `127.0.0.1` and prints a JSON object
containing the port it is listing on and its PID followed by a newline
to standard out.

Clients may connect via WebSockets to one of three services:

- `ws://127.0.0.1:XYZ/EnvironmentService/v1.0`
- `ws://127.0.0.1:XYZ/ServerService/v1.0`
- `ws://127.0.0.1:XYZ/AspireService/v1.0`

Each of these corresponds to a set of services VS Expects (their names
map to the I<Foo>Service.cs interfaces in their source tree)

The general interaction model is that a client connects and calls the
`InitializeAsync` method on the `ServerService` service which gives it
back an opqaue session ID. The client can then call any methods on the
`EnvironmentService` or `AspireService`. Internally the server tracks
these ids to allow caching of data between a single logical
connection. Each session also contains a scoped IoC container which is
used for resolving dependencies, allowing us to leverage much of the
existing `azd` implementation.  We do need to think long term about
lifetime in this new model, but things seem to hang together okay
right now.

Most methods have an `IObserver<ProgressMessage>` parameter that
allows `azd` to write messages during operations back to VS.  This
uses some special features of the `StreamJsonRpc` package, from .NET,
we implement their strategy to allow it to marshal a proxy for the
IObserver to us, which we can then call `onNext` on.

You can use `StreamJsonRpc` (as VS Does) to interact with these
endpoints. A small test program is included in
`vsrpc/testdata/dotnet-azd-client` which exercises these RPCs and
writes log output which can be helpful when debugging the server. It
expects to connect to the server on port 8080, so use the `--port
8080` flag when launching the server.

Fixes #2763