Remove support for `.connection.v0` resource types (#3232)
Aspire removed the `.connection.v0` resource types in Preview 2. We
kept our support in because it was easy to do so and gave folks a
chance to migrate from Preview 1 to Preview 2.

However, as we get ready for Aspire Preview 3, it's time for us to
remove support from `azd`. We'll never see these in manifests anymore,
so it is safe to remvoe the code.