Don't show the spinner when output is not a terminal (#3247)
When the output is not a terminal, we can't erase the spinner, and so
the spinner package just prints the next step of the spinner each time
the message changes. For us that leads to output like:

````
| =     | Doing some long running thing
| =     | Doing some other running thing
`````

This doesn't look great.

Arrange things such that when we are not connected to a terminal, we
just don't have a spinner animation. We still print the message, there
is just no artwork to go with it.