Allow caller to control root container in `NewRootCmd`
This change adds a parameter to `NewRootCmd` to allow a caller to pass
the IoC container that is used during construction of the root cobra
command. Doing so allows a caller to change the result of
registrations (as our vs-server needs to do).

Because `NewRootCmd` registers dependencies itself, a caller needs to
take care to override whatever they want after calling `NewRootCmd`
but before trying to actually run any methods on the returned
`*cobra.Command`.

In an ideal world, `NewRootCommand` would not do any registrations and
the caller would be required to do all the registrations.  We'll get
there next. Right now this change provides the surgical hook needed by
the VS sever work without being overly distruptive to the codebase.