Move `envFlag` out of `cmd` (#3262)
The VS Server work is going to want to reference this type (it needs
to register a provider for it in the IoC container because our actions
pull it when loading an Environment) which means it can't live in the
`cmd` package.

The `internal` package is not a great long term home for this, but it
is where `GlobalCommandOptions` lives today and these two types are
deeply related. Long term I'd like to move these both somewhere else
since bare `internal.Xyz` imports read very poorly to me, but now at
least these two types live in the same package and can move to a their
final home together.