Fix duplicated output for spinner updates in non-tty scenarios (#3257)
The spinner library we're using triggers a paint update when any data update occurs in non-tty scenarios. This unfortunately results in a lot of unexpected extra output in non-tty scenarios.

With this change, we sidestep the issues here by being more careful on how we use the spinner and add tests to verify the desired behavior. Perhaps down the line we could consider addressing this cleanly with our own spinner implementation.
 
Fixes #3254