Improve test coverage in `internal/vsrpc`
This change adds some tests around the infrastructure for serving our
RPC endpoints.

We now have coverage of our logic for cancellation and our support for
`IObserver` as well as coverage of all our `Handler[Action|Func][0-3]`
functions which validate the core marshalling logic of our server.

While writing these tests, I discovered a small functional bug in our
logic for unmarshalling `IObserver`. We intended to validate that the
tag when present was 1, but due to a logic error this was not
happening.