Improve test coverage in `internal/vsrpc`
This change adds coverage for the `GetAspireHostAsync` RPC in
`AspireService` as well as some of the session logic that is common
across all services.

While writing the tests, I realized that it would be both easier and
more correct to use `project.Load` directly in `GetAspireHostAsync`
instead of trying to resolve a `*ProjectConfig` from the IoC
container. This better reflects the intention of the code. In
practice, the production implementation of the resolve function just
called `project.Load` anyway, so there is no functional
difference. However, in cases where the `azure.yaml` is corrupt in
some manner, we'll have better error messages here as well.