Use transient lifetime in `vs-server.go`
This is a quick fix to work around some lifetime issues in vs-server
and the ioc container.  A bug tracks fixing this.  Without this fix,
the end to end test script would crash for an un `init`'d respostiory.

I validated by hand with this fix the script works for both an init
and uninit'd project.