Document external authentication prototcol
This feature was added in #2541, but at the time we did not write
complete documentation for how the feature worked, instead there was
just a good commit message that outlined the work.

We now want to use a similar strategy with VS, so this document
outlines the protocol that we came up with to be able to hand
authentication off to an external process.