Remove session caching and state (#3381)
Removes `container` and `manifest` caching from the session. With this change, each RPC request creates a container copy with only registrations and recreates components fully to handle the particular request. This sidesteps concurrency issues with current component designs that may not be handled fully.

This effectively means all components are "request scoped" and would perform the same app initialization logic as-if the "azd cli" process had started from brand new. This allows for concurrency and avoids blocking, at the cost of memory + I/O redundancy. This is an expedient change to increase reliability and avoid any caching issues. In the future, as we start refactoring the app to separate out different systems that are initialized differently, we can reshare components + implement different I/O polling strategies for the complete e2e to work more efficiently.