vs-server: Include `path` to service instead of `kind` (#3424)
* vs-server: Include `path` to service instead of `kind`

Neither Aspire nor `azd` proper have the concept for a "kind" for a
service. Instead of forcing us to come up with some meaning, simply
return the full path to the project, as we would in other cases like
`azd show`. Consumers can use the path to determine the "kind" of
service it is, using whatever huristics they like.

Fixes #3374

* Update test assertions

---------

Co-authored-by: Wei Lim <weilim@microsoft.com>