vs-server: Support External Auth
This change adds support for clients using the vs-server to provide an
external authentication server (as described in
cli/azd/docs/external-authentication.md) when initializing a
session. When a client does, all auth requests for that session will
use the configured server, allowing the client to fetch a token on
behalf of `azd`.  Visual Studio will use this to request a token using
the current signed in user, instead of forcing users to also run `azd
auth login`.

This configuation is optional - when not provided, `azd` will continue
to use the user signed in via `azd auth login`.

The `InitializeServerAsync` RPC gains a new argument -
`InitializeServerOptions` which allows the values that would normally
be set via `AZD_AUTH_ENDPOINT` and `AZD_AUTH_KEY` environment
variables. The `vs-server` no longer respects these environment variables.

Fixes #3320