vs-server: Validate correct number of arguments for RPCs
Our handler logic for RPCs ensured that the client at least as many
arguments as an RPC required, but any extra arguments from the caller
would be silently ignored.

This change ensures that when a caller passes more arguments than
expected we correctly return an error, as we would if the caller
passed too few.