Consider configured subscription in `TokenCredential` registration.
This change updates our registration of `TokenCredential` in our IoC
container to consider the value of `AZURE_SUBSCRIPTION_ID` when it is
set in the environment. When set, we now fetch the tenant that this
subscription belongs to and use that as the default tenant for auth
requests. This is required for subscriptions which are in a non
default teant for the user.

Fixes: #3485