Don't advertise the unified container in PR builds.
Long term we hope to stop building this container - but we aren't yet
sure who we might impact by doing so, so it will be a gradual effort.

To start, we can at least stop linking to it in our per PR build
message - that way folks are less likely to notice it and take a
dependency on it and if someone points out that it is missing we can
ask them what they are using it for today and advise them on how to
move away from it and onto something we'd like to support long term.

Contributes To: #3494