vs-server: Treat `panic`s in handlers as RPC errors instead of crashing (#3529)
This change modifies the behavior of the RPC server such that if a
handler panics, instead of tearing down the entire process we have the
RPC in question return an RPC error (with the Code set to
`InternalError`).

In theory `azd` does not panic during any of these handlers. In
practice, sometimes we make mistakes an getting an error back instead
of `azd` crashing an the RPC failing when the connection is terminated
is a much nicer experience and also ensures the error message makes
its way back to to the caller.