vs-server: Implement `DeleteEnvironmentAsync` (#3560)
Implement `DeleteEnvironmentAsync`. 

This change adds `Delete` functionality to the `environment` package. `Delete` only deletes the local environment. When the environment is selected the current default environment, it is cleared from the environment configuration.

The `Delete` functionality is surfaced in vs-server as `DeleteEnvironmentAsync`. `DeleteEnvironmentAsync` also supports deleting remote Azure resources, which translates to running `azd down --force --purge` on the environment.

Fixes #3285