Remove `azcore.TokenCredential` from DI graph.
The idea behind this change is avoid "hiding important abstraction". It is important for
API calls to think about: which tenant credential should I be using, in a multi-tenant application.

In most cases, it's derived directly from the subscription. But in some cases, it might require an explicit tenant.
In all cases, it's important to consider this detail in implementation.