Remove `azcore.TokenCredential` from DI graph. (#3638)
With this change, we avoid registering a tenant-specific `azcore.TokenCredential` in the DI graph and fix consuming code using the wrong tenant for their respective use cases. We do this to expose important implementation details of which tenant or subscription should be used to produce an `azcore.TokenCredential` (which by definition, is always tenant-specific). Such a detail is quite important in a multitenant application.

Fixes #3485