Add `--deployment-scope` to `infra synth`
With this change, you may now pass `--deployment-scope resourceGroup`
to `azd infra synth` to have it generate a resource group scoped
deployment for an Aspire application. This causes the generated
`main.bicep` to be scoped to a resource group, not a subscription, and
to not create the resource group that all the aspire resources are
deployed into. Instead the current resource group is reused.