pipeline: Allow configuration of client secret lifetime
With this change, `azd pipeline config` gains a new flag,
`--client-secret-duration` which allows control over how long the
client secret we create when using a shared secret instead of
federated credentails is valid for. This defaults to 180 days, the
previously unconfigurable default.

The value should be a string as recognized by `time.ParseDuration`.