vs-server: Add `--use-tls` for serving over TLS (#3700)
While communication only happens over localhost, we'd still like the
use TLS on the endpoint to prevent other processes from sniffing local
traffic and to allow the client to validate that it is talking to the
server it expects (which it can do by ensuring the certificate
presented during the TLS Handshake matches what is expected).

With this change, `azd vs-server` gains a new parameter,
`--use-tls`. When set, it generates a self-signed 2048 RSA certificate
on each run and uses it to expose its endpoints over TLS. This
certificate is ephemeral and re-generated each time the server is
launched. The lifetime of the generated certificate is short (7 days)
per guidence from the security team. This means the server will need
to be recycled after seven days because the certificate it is using
will no longer be valid, but that should be fine for our current
integrations.

The raw bytes of the certificate are returned in the new
`certificateBytes` property of the JSON object that is printed to
`stdout` by `azd vs-server`. This is a Base 64 encoding of the
certificate.

.NET Consumers may use `new
X509Certificate2(Convert.FromBase64String(...))` to get a .NET object
representing the certificate. Since the certificate is self signed, by
default, ClientWebSocket will fail to connect to the server, so a
consumer must provide a custom `RemoteCertificateValidationCallback`
delegate that validates the certificate and allows it.

This value is not present when TLS is not enabled (so the caller can
also use it's existance to decide if it should use `ws://` or `wss://`
when trying to talk to the server).

Fixes #3684