Enable DataProtection Runtime feature in ACA (#3711)
This change does two things:

- Supports using a custom api-version of the ACA CreateOrUpdate
request by allowing the yaml deployment template to have an
`api-version` property at top level. When set, this value is used as
the `api-version` of the request. The request body is the object,
rendered as JSON, with the `api-version` key removed.

- Uses the above new feature to opt into the `2024-02-02-preview` in
Aspire and sets the
`configuration.runtime.dotnet.autoConfigureDataProtection` property to
`true`.

Fixes #3538