Support `args` on `project.v0` and `dockerfile.v0` resources
We allow binding expressions to refer to properties of other
resources, but we fail if any of these values are secrets (since ACA
has no way to treat these as ACA or KeyVault secrets and reference
them). If we detect this case, we fail and encourage the user to pass
these values via environment variaibles.

Fixes #3678