Support DataProtection Runtime feature in ACA
This change does two things:

- Supports using a custom api-version of the ACA CreateOrUpdate
request by allowing the yaml deployment template to have an
`api-version` property at top level. When set, this value is used as
the `api-version` of the request. The request body is the object,
rendered as JSON, with the `api-version` key removed.

- Uses the above new feature to opt into the `2024-02-02-preview` in
Aspire and sets the
`configuration.runtime.dotnet.autoConfigureDataProtection` property to
`true`, but only when the alpha feature
`aspire.autoConfigureDataProtection` is turned on.

We are gating turning this feature on behind a feature flag the user
must opt-into instead of always sending this request, as we had done
in #3711, because currently regions in Azure are rejecting the
`2024-02-02-preview` requests (even though the error message implies
this should work).

This allows partners to test the end to end by running:

`azd config set alpha.aspire.autoConfigureDataProtection on`

And testing in a supported region, like `centraluseuap` (which can be
explicitly selected via `azd env set AZURE_LOCATION centraluseuap`.

Fixes #3538