Avoid re-prompting for Bicep parameters. (#3734)
This change makes `bicepProvider.Initialize` assert that all client-state is ready for provisioning, failing if any errors occurred. In doing so, we avoid re-prompting for Bicep parameters.

To avoid any behavioral changes in regard to when prompting occurs during `up`, we have moved that logic into `up` explicitly until we decide directionally where it should go in #3745.