aspire: Write all manifests under AppHost folder
Previously when running `azd infra synth` to write out the
infrastructure for an Aspire application, we'd write the container app
manifests into files named `containerApp.tmpl.yaml` in a `manifests`
directory under the project folder for the individual project.

This change adopts a new strategy, now we instead create a single
`infra` folder next to the AppHost project and write the container app
manifest for every `project.v0` and `dockerfile.v0` resource into that
folder, named after their aspire resource name (i.e. the first
argument passed to `AddProject`) with the `.tmpl.yaml` suffix.

This prepares us for a future change where we will move the IaC for
`container.v0` resources out of the `resources.bicep` file and into
individual manifests per resource in this same folder.

It also allows the same project to be reference by two seperate app
hosts and then `infra synth` both of them without having collisions,
or to have the same project (with different configuration) be added to
the App Host multiple times. This isn't something that is commonly
done today but the ability to do this is nice.

Fixes #3761