aspire: Model `container.v0` like `project.v0` and `dockerfile.v0`
This changes the way that we model `container.v0` resources from an
App Host project to be more in line with how `project.v0` and
`dockerfile.v0`. When we started to explore what `azd` and Aspire
looked like, we decided that since we didn't have to worry about
building and pushing container images for a `container.v0` resource,
we could just emit the bicep for the container app as part of the
shared infrastructure for the application and not worry about trying
to create these images at `deploy` time.

As time has gone on, we've found that treating these special was the
wrong call. They are very similar to the `project.v0` and
`dockerfile.v0` resources and modeling them the same way makes the
whole system easier to understand.

With this change, we now treat these resources the same as
`project.v0` and `dockerfile.v0`. This means that we don't emit bicep
for these resources (instead we have `.tmpl.yaml` manifests) and that
we don't create the container app resources until `azd deploy`.

While creating the container app itself has moved into `deploy`, we
still create any volumes needed in the managed environment during
`provision`. The names of these volumes written as bicep outputs
(using `SERVICE_<service_name>_VOLUME_<volume_name>_NAME` as a
template) and then referred to via `.Env` references in our container
app manifests.