dotnet: Harden parser logic in `getTargetPort` (#3833)
In some cases the `dotnet` tool will helpful write a message telling
you that a workload is out of date to standard out when you are using
the new `--getProperty` stuff from msbuild to get JSON output from the
tool.

This was breaking our logic to sniff the target port for a dotnet
container.

In the past we tried to harden this by cleaning the input to the
function, but we found cases in the wild where that didn't seem to
work because our cleaning huristics were not robust.

To address this, rewrite the logic such that we use `json.NewDecoder`
to pull a single JSON value from the start of the stream, validate
it's an object with a config property and then decode it into the
expected format. This allows us to ignore any cruft at the end of the
string.