Updates devcenter config to param plus updates to setting changed values (#3830)
Resolves #3826

Devcenter configuration can exist at multiple levels for azd

Environment variables
Global azd configuration
Project configuration in azure.yaml
azd envrionment configuration within .azure/<env-name>/config.json
Since the config object was being injected as a transient component changes to the config where not getting picked up when changes where made based on user prompts.

This PR removes the config from an injected param of a couple components and into a function param for consumed components. This allows the correct config to flow through to the required functions.

In addition to this this enables us to now only save updates to the config based on what has changed from the originally loaded value as part of the azd environment configuration.