cli: Add "Demo Mode" which hides subscription IDs
When producing content, folks often spend time scrubbing or masking
subscription IDs from the output of `azd`. Instead of folks having to spend their time on this, it would be nice if `azd` could do it itself.

This change adds support for setting an environment variable,
`AZD_DEMO_MODE`, which controls if we will show subscription IDs or
not.

In cases where we'd previously show a subscription name and ID (like
the selection box when we ask a user to pick a subscription, or when a
`provision` starts and we print the target subscription and location
information) we now just show the name.

In cases where we'd show a link to the Portal in messages (e.g. during
a deployment when we provide a link into the deployment status) we now
just elide the link since the portal links contain the subscription ID
as part of the URL.

This saves folks creating content a bunch of time and still keeps the
UX very close to what customers actually see (where the subscription
ID is useful information to have.)

Fixes #319