Key `subscriptions.cache` by user account ID (#3842)
Key `subscriptions.cache` by user account ID.

The cache now stores a map of account ID to subscriptions. By default, account ID is chosen to be the `oid` claim of the user. When `oid` claim is not present (personal accounts), the `sub` (subject) claim is used as the account ID instead. 

This allows VS users to be able to switch between accounts without require running `azd auth logout` in-between.

The cache will be transparently upgraded with this change. Manually tested switching between personal and work accounts in CLI and VS.

Fixes #3654, #2301