Updating service plan to B3.
Templates were using B1 which takes a long time to boot and be ready to deploy.
Also, when using Oryx, it takes more than 3 minutes to build a node app.

The experience from azd deploy is very bad for webApp using B1 and the cost benefit is not that much compared to B3 (.017 v/s .067 p/h).

The fact that B1 runs with only 1 vCPU drastically reduces its performance compared to B3 running with 4 vCPUs