synth: Warn when AppHost has changed since generation
This changes adds a commented to files generated by `azd infra synth`
to include a hash of the manifest that was used to generate the
infrastructure. When we use these files later, we compare the hash
recorded in the file to the hash of the current manifest and issue a
warning.