Sync container app revision secrets (#3877)
Fix KeyVault secrets not being preserved for container app deployments.

We currently do a client-side fetch and patch operation on container app revisions for traditional (non-manifest) container app deployments. This involves fetching and setting the secrets since it's not returned in the initial container app GET request. As such, we need to roundtrip all available information, which includes the KeyVault related properties.

Fixes #3871