dotnet: Use `dotnet publish` to build and push container images
When deploying to a host that requires a container image, if the
`language` of a project was `dotnet` we required the use of Docker to
both build and push the container image. If the user had not authored
a Dockerfile, we would try to use Oryx to build the container image
"from source".

With .NET 8, the SDK itself now has good support for both building a
container image for an app and pushing it to a remote registry. We
leveraged this for our Aspire work but non Aspire dotnet projects
couldn't leverage this.

This change fixes the situation. If your language is `dotnet` and your
host requires a container, we'll now use `dotnet publish
/t:PublishContainer` to build and push the container image.

This does mean that the typical "Build" and "Package" steps behave
differently for these apps now.  Since our "Publish" step now both
builds and pushes the container image, we defer the the work we would
have done in "Build" and "Package" into the "Publish" step. Doing this
allows us to eschew the need for a local Docker daemon to be installed
and running.

The .NET tooling does support creating a `.tgz` which could later be
`docker load`'ed, and you could imagine that "Build" or "Package"
produces such a thing, but doesn't allow us to use this as the input
to a publish, so for now we don't spend time producing this artifact
which would be unused by the downstream publish step.

Fixes #2632