aspire: Support Azure Functions Prototype
The Aspire team has started to prototype Azure Functions support and
this PR is the start of the support in `azd` for this deployment
target under Aspire.

In the current prototype, Aspire models this as a new `function.v0` resource type, which `azd` has special knowledge of. We need to do two special things with this resource type:

1. When generating IaC for the project, we need to generate IaC that
allows us to use the Functions on ACA feature (where we can create a
`Microsoft.Web/sites` resource with correct configuration to run a
functions app on an existing container app environment)

2. Since function projects do not support using `dotnet publish` to
produce a container image (there is work in flight to support this),
we require a `Dockerfile` at the root of the project, which we `docker
build`.

In the next phase of the prototype, we will replace `function.v0` with
`project.v0`, so the `function.v0` manifest type will not live
forever, but until, `dotnet run` and `dotnet publish` works as
expected for Azure Functions projects, we'll continue to model this as
its own type in the manifest.

Note that there are still a lot of rough edges here... The initial
`azd up` works if you ensure your function app name in the app host is
unique across all of `azure` but future updates seem to hang. Also,
none of the corresponding aspire support is in tree yet, so this is
not something that anyone can just pick up and use right now, but I
want to get the work out from behind our `-pr` repository and rebased
on top of the latest main as we are starting to pick up steam on this
work.