Add `--managed-identity` to `azd auth login` (#3723)
We had seen requests for this in the past, e.g., #2568. It would be
useful to have for cases where you are running `azd` inside Azure (say
as part of some deployment system you've put together) and want to use
managed identity instead of having to have a long lived service
princiapl.

By default `--managed-identity` uses the system assigned managed
identity, but `--client-id` can be passed to use a user assigned one.