aspire: Rework image naming strategy
This change adopts the same naming strategy that we use for non aspire
based container workloads where we need to generate an image name.

This uses the same image name (but different tags) instead of creating
a new image (with the timestamp baked into the image name) per deploy.

Fixes #3335