WIP: Support project.v1
This change adds support for the `project.v1` resource we are
designing, which allowes a manifest to include a reference to a bicep
module that should be used when deploying a service, instead of the
IaC that `azd` would generate itself.

Inside the manifest, a `project.v1` resource looks similar to a
`project.v0` resource, with some small changes:

1. The `env` property of the resource is removed, since `azd` no
longer needs to generate the IaC for the service.  2. A new
`deployment` property exists, which is an object that includes
information about what module to use for deployment. This is modeled
after a `azure.bicep.v0` resource

Here's an example, using the `aspire-starter` template, where I have moved the `apiservice` service to be a new `project.v1` resource:

```json
"apiservice": {
  "type": "project.v1",
  "path": "../AspireStarter12.ApiService/AspireStarter12.ApiService.csproj",
  "bindings": {
    "http": {
      "scheme": "http",
      "protocol": "tcp",
      "transport": "http"
    },
    "https": {
      "scheme": "https",
      "protocol": "tcp",
      "transport": "http"
    }
  },
  "deployment": {
    "type": "azure.bicep.v0",
    "path": "apiservice.bicep",
    "params": {
      "containerAppEnvironmentId": "",
      "containerRegistryEndpoint": "",
      "managedIdentityClientId": "",
      "managedIdentityId": "",
      "image": "",
      "targetPort": ""
    }
  }
}
```

Note that today the only supported `type` for a `deployment` is
`azure.bicep.v0`, but this would allow us to support other styles of
deployment manifests (e.g k8s yaml's) down the line.  The `params`
array represents parameters that will be needed to be passed to the
module as part of deployment.  In this mode, we now support `image`
and `targetPort` as two new properties that you can have in this
`params` array, which `azd` will auto fill in (this matches the
bahavior of things like `containerAppEnvironmentId` which is already
an `azd` intrinsic.