aspire: Fix panic when a project has no endpoints
Worker projects do not expose endpoints and hence do not need ingress
configured on their container apps. We had a bug in constructing the
template context in the case where a project did not have and
endpoints.

To address this, we only set `TargetPortExpression` in the case where
we have Ingress configured. `TargetPortExpression` is only used in the
template when the service has ingress, so the empty value will cause
no harm in cases where the service has not configured ingress.

Fixes #4064