async: Remove `InteractiveTaskWithProgress`
This abstraction was not in use in the codebase anymore. We've found
that we are not sold on the entire `Task` abstraction right now (in
practice we don't end up using much that the abstraction buys us and
the way we use it adds friction to our interactive debugging story
because you can't easily "step into" a Task.Run since it spawns a
goroutine to run the user's code) and we'd like to consider exploring
other options.

In the mean time, since this version is un-used, let's just remove it,
to make it a little harder to take a accidental dependency on a
pattern we aren't fond of. If we end up needing something like this,
we can always revive it from the git history.