provisioning: Don't "import dot" the provisioning package
Go has a feature where you can import a package in a way such that you
don't have to qualify symbols with the package name, buy using `.` as
the import name.

We used this in the provisioning sub-packages to import `provisioning`
such that it "felt like" it was part of the sub-package.

I think this process in practice leads to more confusion instead of
less, it becomes less obvious about where things are defined and what
is depending on what and how.

The Go Wiki also discourages the use of this pattern except in
cases in tests where you need it to address a circular dependency
issue, which we do not face here.

This change removes the use of the pattern across `azd` (we only used
it in the provisioning package, and I think this is an artifact of how
it evolved when we introduced the terraform provider to be a sibling
of the exisitng bicep provider (which birthed the provisioning
interface).