aspire: Remove support for legacy resources (#4021)
During early previews of Aspire, there was a large set of custom
resource types that Aspire and `azd` supported. As the product
evolved, these types were removed from the manifest in favor of a
single `azure.bicep.v0` type, but we retained our support for the
older stuff as a point in time thing.

That point in time has now passed, Aspire has GA'd without any of
these resource types in it. We can now remove our support for them and
remove a bunch of logically dead code.

Fixes #4906