Remove `async.Task[T]` and friends (#4122)
This change removes the use of `async.Task` in favor of writing
functions in a more idiomatic go style.  We introduced this
abstraction a while back to provide some structure around our use of
channels and to model operations in a way like `Task<T>` from .NET or
`Promise<T>` from JS, but I don't think it's pulling its weight.

There are a few bothersome issues with `async.Task` today:

1. The use of `SetError` and `SetResult` does not feel idiomatic to go
where normal multi-value returns are used. Calling `SetError` or
`SetResult` must always be followed by a call to `return`.  We've had
issues in the past where we haven't called `return` in code like this:

```go
res, err := foo()
if err != nil {
    task.SetError(err)
    // oops, forgot to return, so we continue and likely crash when we try to touch res.
}
```

2. The use of goroutines makes the code difficult to debug. Calls to
`Task.Run` with a delegate end up invoking work on a new goroutine,
which means it's more difficult to "step into" the function in the
debugger. With a normal go function, this is not the case.

3. When one async function calls another and they both have progress,
you must manually wire up writing events from one progress stream to
another (we had a helper, `syncProgress` here, but you had to call it
after calling the async function but before awaiting).

4. Our use of how we structured functions and used `Task.Run` meant
that in practice we had two extra levels of indentation is a bunch of
our code, which made it harder to read.

To remove `async.Task` I have introduced `async.Progress[P]` which is
just a wrapper around a `chan P`. This is logically the `WithProgress`
part of `TaskWithProgress` but on its own type.  This is now passed as
an argument to any function which used to return a `TaskWithProgress`.
If one function that reports progress is called by another, it can now
pass the progress instance into the callee and not have to worry about
the `syncProgress` dance.  With `progress` decoupled from the task
context, it was then possible to just convert all the functions to be
normal go style functions that return a result and an error.

Compared with the previous implementation, it is important that when a
`Progress` is created, a reader of it is started before calling a
function which will write to it (since the call to `SetProgress` will
block until a reader picks up the value).  This is backwards from the
old pattern where you would run the function to create the task, then
start to observe the progress and then call await.  You can use
`async.RunWithProgress` to manage much of the ceremony for you.

Turning on "Ignore whitespace" is likely helpful when reviewing this,
since a lot of the code just ended up getting outdented as we removed
levels of `Task.Run` but didn't actually change.

Fixes #2354